Terminate child on exit timeout
Previously, when ghostunnel would hit the graceful shutdown timeout,
it would simply exit without signaling to the child process. This
changes the logic to make sure we always terminate the child on exit.

In addition, we also add a test to make sure that a misbehaving child
which is ignoring SIGTERM is properly killed after timeout.