Improve keychain/certstore certificate choice
It's possible for the macOS keychain to have multiple certificates with
the same name. In this case, the code used to pick whichever one
arbitrarily showed up first in the certstore.Identities() results. This
tended to be the first one added, which is usually not the most up to
date one. Instead, this now will pick the one with the most distant
NotAfter date.