Introduce TLSConfigSource interface
The existing Certificate interface is not sufficient in situations where
more control is needed over the TLS configuration beyond the
certificates and CAs used in the handshake.

This commit introduces a new interface, the TLSConfigSource, that is
used to obtain server or client TLS configuration and allows an
implementor to have total control over the TLS configuration.

This change is a precursor to SPIFFE Workload API and SPIFFE
authentication support within ghostunnel, which specifically needs
to expand/override the peer certificate verification code.

A wrapper is provided that implements a TLSConfigSource in terms of the
existing Certificate interface to minimize disruption to the code base.