Relax curve prefs in client mode; fixes issue #232
When in server mode, we set curve preference to make sure clients
won't be able to select inefficiently implemented curves, at least
for x86-64 platforms. But in client mode that argument doesn't hold,
and makes ghostunnel clients incompatible with some backend services
where we don't control their curve prefs. This splits the template
config for client/server so we can relax the curve prefs for clients.