test-keys: fix order of prerequisites in rules
The change introduced in the commit `396aebe83bf353e2f` broke the build
rules in the recipes, as the latter depend heavily, but not explicitly,
in the order of the prerequisites.  Thus, it is very easy for someone to
break these rules.

To make the order more explicit, I added a short notice.  Ideally, we
should not depend on such order or let make(1) figure out automatically
the correct elements.