Add initial ACME support in server mode (#348)
* Add initial ACME support in server mode

Add basic support for using ACME and Let's Encrypt to obtain a public
TLS certificate. This is intentionally limited to server mode. Only
the TLS-ALPN challenge type is supported.

* Resolve suggestions from linter

* Updates to ACME Support

* Allow the user to specify the URLs for a production ACME CA
and a test ACME CA. If not specified, default to Let's Encrypt.

* Add a mandatory command line flag explicitly capturing that the user
accepts the terms of service of the ACME CA.