Add optional /_shutdown endpoint for terminating via HTTP POST
When running as a sidecar container in Kubernetes to jobs, the primary
container needs a way to signal to the other containers they should also
terminate. Although you can share the process namespace, this is simpler
to achieve since you don't have to look for the correct PID.

I also tweaked the sigterm tests to only TERM once and without an open
connection since while working on the HTTP tests (which were modeled after
those) the server was reaching the timeout because of the open connection
which is already covered in the timeout tests.