Better landlock rule processing
Some flags can contain both a host:port address as well as a URL
starting with http(s)://. For example, the status flag can contain a
host:port to serve status on HTTPS by default. But it can also
explicitly contain an HTTP URL to disable HTTPS.

To accommodate flags containing both possible values, we combine the
parsing into one function and look at the http(s):// prefix to see if
we're dealing with a URL instead of a host:port.